refactor api for evaluation to have better codewise
